Feat: 로그인 토큰을 얻기 위해 임시로 로그인 함수 생성, response 타입 추가
