+++ pandamarket-ts/types/articleResponseTypes.ts
@@ -32,19 +32,4 @@ export interface ArticleCommentsResponse {
   list: Comment[];
 }
 
-export interface SignInResponse {
-  refreshToken: string;
-  accessToken: string;
-  user: {
-    updatedAt: string;
-    createdAt: string;
-    image?: {
-      UrlType: string;
-    };
-    nickname: string;
-    id: number;
-    email: string;
-  };
-}
-
 export type OrderBy = "like" | "recent";
